refactor(ui): extract health/components.ts — HealthCard, HealthActionRow, StatBlock + render helpers
Move the Preact component primitives and their render helpers out of
health.ts into health/components.ts: HealthCard, HealthActionRow,
StatBlock, buildHealthCard, renderStatBlocks, renderText, renderIcons,
renderHealthCards, renderActionList, and the private formatStatValue
helper.
